+++ src/main/java/de/bixilon/minosoft/gui/rendering/gui/elements/text/mark/MarkTextElement.kt
@@ -77,15 +77,15 @@ class MarkTextElement(
             val message = chatComponent.message // ToDo: This does not include formatting
             val preMark = TextElement(guiRenderer, message.substring(0, markStartPosition), properties = properties, parent = _parent)
             val mark = TextElement(guiRenderer, message.substring(markStartPosition, markEndPosition), properties = properties, parent = _parent)
-            val markOffset = Vec2i(preMark.info.lines.lastOrNull()?.width ?: 0, preMark.size.y)
+            val markOffset = Vec2f(preMark.info.lines.lastOrNull()?.width ?: 0f, preMark.size.y)
             if (markOffset.y > 0 && (preMark.info.lines.lastOrNull()?.width ?: 0.0f) <= (info.lines.lastOrNull()?.width ?: 0.0f)) {
                 markOffset.y -= (properties.lineHeight).toInt()
             }
 
             for (line in mark.info.lines) {
-                ColorElement(guiRenderer, size = Vec2f(line.width, (properties.lineHeight).toInt()), color = ChatColors.DARK_BLUE).render(offset + markOffset, consumer, options)
+                ColorElement(guiRenderer, size = Vec2f(line.width, properties.lineHeight), color = ChatColors.DARK_BLUE).render(offset + markOffset, consumer, options)
                 markOffset.x = 0
-                markOffset.y += (properties.lineHeight).toInt()
+                markOffset.y += properties.lineHeight
             }
         }
